PR change: util function + shift dropdown back to grading page
